Move Span to candidates phase
