+++ fonduer/candidates/models/__init__.py
@@ -1,12 +1,15 @@
 from fonduer.candidates.models.candidate import Candidate, candidate_subclass
-from fonduer.candidates.models.implicitspan import ImplicitSpan, TemporaryImplicitSpan
 from fonduer.candidates.models.image import Image, TemporaryImage
+from fonduer.candidates.models.implicitspan import ImplicitSpan, TemporaryImplicitSpan
+from fonduer.candidates.models.span import Span, TemporarySpan
 
 __all__ = [
     "Candidate",
     "Image",
     "ImplicitSpan",
+    "Span",
     "TemporaryImage",
     "TemporaryImplicitSpan",
+    "TemporarySpan",
     "candidate_subclass",
 ]
