+++ fonduer/candidates/models/span.py
