+++ fonduer/parser/models/__init__.py
@@ -2,10 +2,6 @@ from fonduer.parser.models.context import Context, construct_stable_id, split_st
 from fonduer.parser.models.document import Document
 from fonduer.parser.models.figure import Figure
 from fonduer.parser.models.phrase import Phrase
-from fonduer.parser.models.span import (
-    Span,
-    TemporarySpan,
-)
 from fonduer.parser.models.table import Cell, Table
 from fonduer.parser.models.webpage import Webpage
 
@@ -15,9 +11,7 @@ __all__ = [
     "Document",
     "Figure",
     "Phrase",
-    "Span",
     "Table",
-    "TemporarySpan",
     "Webpage",
     "construct_stable_id",
     "split_stable_id",
