+++ fonduer/supervision/lf_helpers.py
@@ -4,11 +4,9 @@ from collections import defaultdict
 from itertools import chain
 
 import numpy as np
-from lxml import etree
-from lxml.html import fromstring
-
 from fonduer.candidates import Ngrams
-from fonduer.parser.models import Phrase, TemporarySpan
+from fonduer.candidates.models import TemporarySpan
+from fonduer.parser.models import Phrase
 from fonduer.utils import tokens_to_ngrams
 from fonduer.utils.utils_table import (
     is_axis_aligned,
@@ -26,6 +24,8 @@ from fonduer.utils.utils_visual import (
     bbox_vert_aligned_left,
     bbox_vert_aligned_right,
 )
+from lxml import etree
+from lxml.html import fromstring
 
 
 def get_between_ngrams(c, attrib="words", n_min=1, n_max=1, lower=True):
